change all float to doubles to avoid float precision errors

+++ edge.cpp
@@ -5,7 +5,7 @@ edge::edge() {
     next_line = 0;
 }
 
-void edge::addLineWeights(array<float,NBR_CLASSES> weightLine) {
+void edge::addLineWeights(array<double,NBR_CLASSES> weightLine) {
     weights[next_line] = weightLine;
     ++next_line;
 }
