+++ tree.h
@@ -13,11 +13,11 @@ public:
     vector<edge> edges;
 
     tree(vector<node> nodes, vector<edge> edges);
-    float forward_backward_min_marginals();
+    double forward_backward_min_marginals();
     void forward(int until_nodeId);
     void backward(int until_nodeId);
     bool proper_min_marginals();
-    float get_min_marginal_for_id(int node_id);
+    double get_min_marginal_for_id(int node_id);
 };
